put service logic in separate classes and renamed directories
