--- src/main/java/core/basesyntax/dao/InfoFromFileDao.java
@@ -1,9 +1,0 @@
-package core.basesyntax.dao;
-
-import core.basesyntax.model.FruitTransaction;
-
-import java.util.List;
-
-public interface InfoFromFileDao {
-    List<FruitTransaction> getAllTransactionsFromFile(String fileName);
-}
