+++ src/main/java/core/basesyntax/db/Storage.java
@@ -1,0 +1,9 @@
+package core.basesyntax.db;
+
+import core.basesyntax.model.FruitTransaction;
+
+import java.util.List;
+
+public class Storage {
+    public static List<FruitTransaction> fruitTransactions;
+}
