+++ src/main/java/core/basesyntax/model/FruitTransaction.java
@@ -41,7 +41,7 @@ public class FruitTransaction {
         PURCHASE("p"),
         RETURN("r");
 
-        private String code;
+        private final String code;
 
         Operation(String code) {
             this.code = code;
