+++ src/main/java/core/basesyntax/service/CsvFileReaderService.java
@@ -1,0 +1,5 @@
+package core.basesyntax.service;
+
+public interface CsvFileReaderService {
+    String read(String fileName);
+}
