+++ src/main/java/core/basesyntax/service/CsvFileWriterService.java
@@ -1,0 +1,5 @@
+package core.basesyntax.service;
+
+public interface CsvFileWriterService {
+    void write(String string);
+}
