--- src/main/java/core/basesyntax/service/FruitTransactionService.java
@@ -1,7 +1,0 @@
-package core.basesyntax.service;
-
-import java.util.Map;
-
-public interface FruitTransactionService {
-    Map<String, Integer> getReport(String fileName);
-}
