+++ src/main/java/core/basesyntax/service/ReportCreatorService.java
@@ -1,0 +1,7 @@
+package core.basesyntax.service;
+
+import java.util.List;
+
+public interface ReportCreatorService<T> {
+    String create(List<T> info);
+}
