+++ src/main/java/core/basesyntax/service/StringToListService.java
@@ -1,0 +1,7 @@
+package core.basesyntax.service;
+
+import java.util.List;
+
+public interface StringToListService<T> {
+    List<T> convert(String string);
+}
