+++ src/main/resources/data.csv
@@ -5,4 +5,4 @@ p,banana,13
 r,apple,10
 p,apple,20
 p,banana,5
-s,banana,50
+s,banana,1
